# Faster PoT verification without allocation or `std`