--- a/crates/subspace-proof-of-time/Cargo.toml
+++ b/crates/subspace-proof-of-time/Cargo.toml
@@ -3,7 +3,7 @@ name = "subspace-proof-of-time"
 description = "Subspace proof of time implementation"
 license = "0BSD"
 version = "0.1.0"
-authors = ["Rahul Subramaniyam <rahulksnv@gmail.com>"]
+authors = ["Nazar Mokrynskyi <nazar@mokrynskyi.com>"]
 edition = "2021"
 include = [
     "/src",
@@ -19,13 +19,14 @@ aes.workspace = true
 subspace-core-primitives.workspace = true
 thiserror.workspace = true
 
-# This is required to for benchmark dependency features to work correctly
-rand = { workspace = true, optional = true }
+[target.'cfg(target_arch = "x86_64")'.dependencies]
+cpufeatures = { workspace = true }
 
 [dev-dependencies]
 core_affinity.workspace = true
 criterion.workspace = true
-rand.workspace = true
+rand_core = { workspace = true }
+rand_chacha = { workspace = true }
 
 [[bench]]
 name = "pot"
@@ -34,12 +35,3 @@ harness = false
 [[bench]]
 name = "pot-compare-cpu-cores"
 harness = false
-
-[features]
-default = ["std"]
-std = [
-    "subspace-core-primitives/std",
-    "thiserror/std",
-    "rand?/std",
-    "rand?/std_rng",
-]