--- a/crates/subspace-proof-of-time/benches/pot-compare-cpu-cores.rs
+++ b/crates/subspace-proof-of-time/benches/pot-compare-cpu-cores.rs
@@ -1,12 +1,15 @@
 use core::num::NonZeroU32;
-use criterion::{black_box, criterion_group, criterion_main, Criterion};
-use rand::{thread_rng, Rng};
+use criterion::{criterion_group, criterion_main, Criterion};
+use rand_chacha::ChaCha8Rng;
+use rand_core::{RngCore, SeedableRng};
+use std::hint::black_box;
 use subspace_core_primitives::pot::PotSeed;
 use subspace_proof_of_time::prove;
 
 fn criterion_benchmark(c: &mut Criterion) {
+    let mut rng = ChaCha8Rng::from_seed(Default::default());
     let mut seed = PotSeed::default();
-    thread_rng().fill(seed.as_mut());
+    rng.fill_bytes(seed.as_mut());
     // About 1s on 6.0 GHz Raptor Lake CPU (14900K)
     let pot_iterations = NonZeroU32::new(200_032_000).expect("Not zero; qed");