PoT verification is a big part of the block verification cost, especially when syncing otherwise mostly empty blocks.


Before:
```
verify                  time:   [220.72 ms 225.27 ms 229.55 ms]
```
After:
```
verify                  time:   [204.92 ms 205.10 ms 205.30 ms]
```

Tested on Zen 4 CPU.

This change avoids `std` for AES-NI CPU feature detection on x86-64 and avoids heap allocation during verification, allowing for better data locality and overall better cache utilization, while also doing physically less work.

I also changed crate authorship since it was almost completely rewritten by me since its introduction in https://github.com/autonomys/subspace/pull/1601, only some of the tests are still mostly the same.

### Code contributor checklist:
* [x] I have read, understood and followed [contributing guide](https://github.com/autonomys/subspace/blob/main/CONTRIBUTING.md)